app/requirements.txt
requests==2.31.0
click==8.1.3
psycopg2-binary==2.9.6
prettytable==3.8.0